refactor(arch): split apply.ts god file into 4 focused modules
- src/lib/diff.ts: extract computeLineDiff (LCS) + printDryRunDiff
- src/actions/mcp-runtime.ts: extract installLocalMcpRuntime + ensureGitignoreEntry
- src/actions/apply-prune.ts: extract ProtectConfig + runPruneAndProtect loop
- src/actions/apply-output.ts: extract all print* functions + autoInstallSuperpowers
- src/actions/apply.ts: rewritten as pure orchestrator (~345 lines, was 1123)
- eslint.config.mjs: add new action files to no-console exemption, tighten max-lines 500->400
